update swagger docs
