+++ src/routes/token/historical/balances/evm.ts
@@ -2,20 +2,23 @@ import { Hono } from 'hono';
 import { describeRoute } from 'hono-openapi';
 import { resolver, validator } from 'hono-openapi/zod';
 import { handleUsageQueryError, makeUsageQueryJson } from '../../../../handleQuery.js';
-import { evmAddressSchema, statisticsSchema, paginationQuery, intervalSchema, timestampSchema, networkIdSchema, USDC_WETH } from '../../../../types/zod.js';
+import { evmAddressSchema, paginationQuery, statisticsSchema, Vitalik, networkIdSchema, ageSchema, intervalSchema, timestampSchema } from '../../../../types/zod.js';
 import { sqlQueries } from '../../../../sql/index.js';
 import { z } from 'zod';
-import { config } from '../../../../config.js';
+import { config, DEFAULT_AGE } from '../../../../config.js';
+import { injectSymbol } from '../../../../inject/symbol.js';
+import { injectPrices } from '../../../../inject/prices.js';
 
 const route = new Hono();
 
 const paramSchema = z.object({
-    pool: USDC_WETH
+    address: Vitalik,
 });
 
 const querySchema = z.object({
-    network_id: z.optional(networkIdSchema),
     interval: intervalSchema,
+    network_id: z.optional(networkIdSchema),
+    contracts: z.optional(z.string().array()),
     startTime: z.optional(timestampSchema),
     endTime: z.optional(timestampSchema)
 }).merge(paginationQuery);
@@ -23,22 +26,25 @@ const querySchema = z.object({
 const responseSchema = z.object({
     data: z.array(z.object({
         datetime: z.string().datetime(),
-        ticker: z.string(),
+        contract: z.string(),
+        name: z.string(),
+        symbol: z.string(),
+        decimals: z.string(),
         open: z.number(),
         high: z.number(),
         low: z.number(),
         close: z.number(),
-        volume: z.number(),
-        uaw: z.number(),
-        transactions: z.number()
+        // uaw: z.number(),
+        // transactions: z.number()
     })),
     statistics: z.optional(statisticsSchema),
 });
 
 const openapi = describeRoute({
-    summary: 'OHLCV by Pool',
-    description: 'Provides pricing data in the Open/High/Low/Close/Volume (OHCLV) format.',
+    summary: 'Historical Balances',
+    description: 'Provides historical ERC-20 & Native balances by wallet address.',
     tags: ['EVM'],
+    "x-tagGroups": ["Historical"],
     security: [{ bearerAuth: [] }],
     responses: {
         200: {
@@ -48,15 +54,16 @@ const openapi = describeRoute({
                     schema: resolver(responseSchema), example: {
                         data: [
                             {
-                                "datetime": "2025-04-22 01:00:00",
-                                "ticker": "WETHUSDC",
-                                "open": 1568.1231985920829,
-                                "high": 1588.0055525937228,
-                                "low": 1567.4961060936091,
-                                "close": 1587.846767916931,
-                                "volume": 7313447103818,
-                                "uaw": 53,
-                                "transactions": 309
+                                "block_num": 21764208,
+                                "datetime": "2025-02-03 06:31:23",
+                                "date": "2025-02-03",
+                                "contract": "0xc944e90c64b2c07662a292be6244bdf05cda44a7",
+                                "amount": "339640316263000000000000000",
+                                "decimals": 18,
+                                "symbol": "GRT",
+                                "network_id": "mainnet",
+                                "price_usd": 0.10426804866144047,
+                                "value_usd": 35413633.023497514
                             }
                         ]
                     }
@@ -66,18 +73,22 @@ const openapi = describeRoute({
     },
 });
 
-route.get('/:pool', openapi, validator('param', paramSchema), validator('query', querySchema), async (c) => {
-    const parsePool = evmAddressSchema.safeParse(c.req.param("pool"));
-    if (!parsePool.success) return c.json({ error: `Invalid EVM pool: ${parsePool.error.message}` }, 400);
+route.get('/historical/:address', openapi, validator('param', paramSchema), validator('query', querySchema), async (c) => {
+    const parseAddress = evmAddressSchema.safeParse(c.req.param("address"));
+    if (!parseAddress.success) return c.json({ error: `Invalid EVM address: ${parseAddress.error.message}` }, 400);
 
-    const pool = parsePool.data;
+    const parseContracts = z.string().array().optional().safeParse(c.req.queries("contracts"));
+    if (!parseContracts.success) return c.json({ error: `Invalid EVM contracts: ${parseContracts.error.message}` }, 400);
+
+    const address = parseAddress.data;
     const network_id = networkIdSchema.safeParse(c.req.query("network_id")).data ?? config.defaultNetwork;
+    const contracts = parseContracts.data ?? [];
     const database = `${network_id}:${config.dbEvmSuffix}`;
 
-    const query = sqlQueries['ohlcv_prices_for_pool']?.['evm']; // TODO: Load different chain_type queries based on network_id
-    if (!query) return c.json({ error: 'Query for OHLCV pool prices could not be loaded' }, 500);
+    const query = sqlQueries['historical_balances_for_account']?.['evm']; // TODO: Load different chain_type queries based on network_id
+    if (!query) return c.json({ error: 'Query for balances could not be loaded' }, 500);
 
-    const parseIntervalMinute = intervalSchema.transform((interval) => {
+    const parseIntervalMinute = intervalSchema.transform((interval: string) => {
         switch (interval) {
             case '1h':
                 return 60;
@@ -105,14 +116,14 @@ route.get('/:pool', openapi, validator('param', paramSchema), validator('query',
 
     const response = await makeUsageQueryJson(c, [query], {
         network_id,
-        interval_minute: parseIntervalMinute.data,
-        high_quantile: 0.95,
-        low_quantile: 0.05,
-        pool,
+        interval_minute: String(parseIntervalMinute.data),
+        address,
+        contracts,
         min_datetime,
         max_datetime,
     }, { database });
-
+    injectSymbol(response, network_id);
+    // await injectPrices(response, network_id);
     return handleUsageQueryError(c, response);
 });
