+++ src/routes/token/historical/balances/index.ts
@@ -1,0 +1,8 @@
+import { Hono } from 'hono'
+import evm from "./evm.js";
+
+const router = new Hono()
+
+router.route('/evm', evm)
+
+export default router
