+++ src/routes/token/historical/index.ts
@@ -1,0 +1,8 @@
+import { Hono } from 'hono';
+import balances from "./balances/index.js";
+
+const router = new Hono();
+
+router.route('/balances', balances);
+
+export default router;
